third_party/blink/web_tests/clipboard/async-clipboard/text-write-read/async-writeText-readText.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async Clipboard writeText -> readText tests</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../http/tests/resources/permissions-helper.js"></script>
<script>
async function readWriteTest(textInput) {
  await PermissionsHelper.setPermission('clipboard-read', 'granted');
  await PermissionsHelper.setPermission('clipboard-write', 'granted');

  promise_test(async t => {
    await navigator.clipboard.writeText(textInput);
    const textOutput = await navigator.clipboard.readText();

    assert_equals(textOutput, textInput);
  }, 'Verify write and read clipboard given text: ' + textInput);
}

readWriteTest('Clipboard write text -> read text test');
readWriteTest('non-Latin1 text encoding test データ');
</script>
